added

+++ database/migrations/2025_07_13_040000_create_attendance_map_locations_pivot.php
@@ -12,7 +12,7 @@ return new class extends Migration
         // Create pivot table
         Schema::create('attendance_map_locations', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('attendance_id');
+            $table->unsignedBigInteger('attendance_id');
             $table->unsignedInteger('map_location_id');
             $table->timestamps();
